order/requirements.txt
Flask==3.0.2
redis==5.0.3
gunicorn==21.2.0
requests==2.31.0
grpcio==1.70.0
grpcio-tools==1.70.0
pyignite==0.6.1
dotenv==0.9.9
